# target/classes/META-INF/swagger/micronaut-project-1.0.0.yml
openapi: 3.0.1
info:
  title: Micronaut project
  description: "{my.api.description}"
  license:
    name: MIT
  version: 1.0.0
paths:
  /admin/products:
    post:
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "201":
          description: createProduct 201 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: updatedProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProduct"
        required: true
      responses:
        "200":
          description: updatedProduct 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: deleteProduct 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /hello:
    get:
      operationId: helloworld
      responses:
        "200":
          description: helloworld 200 response
          content:
            text/plain:
              schema:
                type: string
  /products:
    get:
      operationId: getproducts
      responses:
        "200":
          description: getproducts 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/filter:
    get:
      operationId: getlistbyQuery
      parameters:
      - name: max
        in: query
        explode: false
        schema:
          type: integer
          format: int32
          nullable: true
      - name: offset
        in: query
        explode: false
        schema:
          type: integer
          format: int32
          nullable: true
      responses:
        "200":
          description: getlistbyQuery 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/type/{type}:
    get:
      operationId: getproductbyType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getproductbyType 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getproductbyid
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: getproductbyid 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /users:
    post:
      operationId: createUserAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: createUserAccount 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
    Product:
      required:
      - id
      - name
      - type
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        type:
          $ref: "#/components/schemas/Product.Type"
    Product.Type:
      type: string
      enum:
      - COFFEE
      - TEA
      - OTHERS
    UpdateProduct:
      required:
      - name
      - type
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/Product.Type"
    User:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        firstName:
          maxLength: 10
          minLength: 3
          type: string
        lastName:
          maxLength: 10
          minLength: 3
          type: string
        email:
          minLength: 1
          type: string
          format: email
        id:
          type: string